Move database logic out of index
The contents have grown enough to warrant its own file instead of
cluttering up the index.
